Tree Search
Create tree search and test for it. Run it a couple times. Find out, that in 90% my results are correct. It is awesome, but I need 100%. Plus my code is running slow, approximately 5 seconds. It is 4 secs to create tree and 1 for search. I actualy thinking about making them in one method. Ok I am really glad, that my project is moving forward, but still a lot of work need to be done.